07/09/16

+++ app/src/main/java/com/progresive/user/cashpokerprogresive/ClaseApuestaPremio.java
@@ -90,9 +90,9 @@ public class ClaseApuestaPremio {
         }
         ApuestaPremioTV.setBackgroundResource(R.drawable.botonapuesta);
     }
-    public void Movimientopremio(int finX,int finY){
-        ObjectAnimator animacionX= ObjectAnimator.ofFloat(ApuestaPremioTV,View.TRANSLATION_X,0f,(float)finX);
-        ObjectAnimator animacionY= ObjectAnimator.ofFloat(ApuestaPremioTV,View.TRANSLATION_Y,0f,(float)finY);
+    public void Movimientopremio(float finX,float finY){
+        ObjectAnimator animacionX= ObjectAnimator.ofFloat(ApuestaPremioTV,View.TRANSLATION_X,0f,finX);
+        ObjectAnimator animacionY= ObjectAnimator.ofFloat(ApuestaPremioTV,View.TRANSLATION_Y,0f,finY);
         ObjectAnimator Agrandadorx = ObjectAnimator.ofFloat(ApuestaPremioTV, View.SCALE_X, 1.3f);
         ObjectAnimator Agrandadory = ObjectAnimator.ofFloat(ApuestaPremioTV, View.SCALE_Y, 1.3f);
         AnimatorSet conjunto = new AnimatorSet();
@@ -101,9 +101,9 @@ public class ClaseApuestaPremio {
         conjunto.setDuration(1000);
         conjunto.start();
     }
-    public void Movimientoapuesta(int finX,int finY){
-        ObjectAnimator animacionX= ObjectAnimator.ofFloat(ApuestaPremioTV,View.TRANSLATION_X,0f,(float)finX);
-        ObjectAnimator animacionY= ObjectAnimator.ofFloat(ApuestaPremioTV,View.TRANSLATION_Y,0f,(float)finY);
+    public void Movimientoapuesta(Float finX,float finY){
+        ObjectAnimator animacionX= ObjectAnimator.ofFloat(ApuestaPremioTV,View.TRANSLATION_X,0f,finX);
+        ObjectAnimator animacionY= ObjectAnimator.ofFloat(ApuestaPremioTV,View.TRANSLATION_Y,0f,finY);
         ObjectAnimator Agrandadorx = ObjectAnimator.ofFloat(ApuestaPremioTV, View.SCALE_X, 1f);
         ObjectAnimator Agrandadory = ObjectAnimator.ofFloat(ApuestaPremioTV, View.SCALE_Y, 1f);
         AnimatorSet conjunto = new AnimatorSet();
